+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Changelog
 * Bugfix: Added `OAuth1ResourceOwner` & `OAuth2ResourceOwner` to cover case of implementing custom oauth resource owners,
 * Bugfix: Fixed Authorization Header in `CleverResourceOwner::doGetRequest`,
 * Bugfix: Catch also the `TransportExceptionInterface` in `AbstractResourceOwner::getResponseContent()` method,
+* Bugfix: Current matched Firewall is respected during generation of resource owner check path links.
 
 ## 2.0.0-BETA2 (2022-01-16)
 * Deprecated: configuration parameter `firewall_names`, firewalls are now computed automatically - all firewalls that have defined `oauth` authenticator/provider will be collected,